+++ balaio/tests/test_validator.py
@@ -1,12 +1,8 @@
 import mocker
+import unittest
 from xml.etree.ElementTree import ElementTree
-import json
 from StringIO import StringIO
 
-from balaio import validator
-from balaio import notifier
-
-class FundingCheckingPipeTest(mocker.MockerTestCase):
 
 class FundingCheckingPipeTest(mocker.MockerTestCase):
 
@@ -68,3 +64,50 @@ class FundingCheckingPipeTest(mocker.MockerTestCase):
         expected = ['ok', '<funding-group>funding data</funding-group>']
 
         self.assertEquals(expected, self._validate('<root><ack>acknowledgements<funding-group>funding data</funding-group></ack></root>'))
+
+
+class ISSNCheckingPipeTest(unittest.TestCase):
+
+    def _make_pipe(self, *args, **kwargs):
+        from balaio.validator import ISSNCheckingPipe
+        return ISSNCheckingPipe(*args, **kwargs)
+
+    def _make_data(self, xml_string='<root></root>'):
+        from StringIO import StringIO
+        etree = ElementTree()
+        return etree.parse(StringIO(xml_string))
+
+    def _validate(self, xml_string):
+        data = self._make_data(xml_string)
+        pipe = self._make_pipe(data)
+        return pipe.validate(data)
+
+    def test_pipe_issn_with_one_valid_ISSN(self):
+        expected = ['ok', '']
+
+        self.assertEquals(
+            self._validate("<root><issn pub-type='epub'>0102-6720</issn></root>"), expected)
+
+    def test_pipe_issn_with_one_invalid_ISSN(self):
+        expected = ['e', 'neither eletronic ISSN nor print ISSN are valid']
+
+        self.assertEquals(
+            self._validate("<root><issn pub-type='ppub'>1234-1234</issn></root>"), expected)
+
+    def test_pipe_issn_with_two_valid_ISSN_eletronic_and_print(self):
+        expected = ['ok', '']
+
+        self.assertEquals(
+            self._validate("<root><issn pub-type='ppub'>0100-879X</issn><issn pub-type='epub'>1414-431X</issn></root>"), expected)
+
+    def test_pipe_issn_with_strange_ISSN(self):
+        expected = ['e', 'neither eletronic ISSN nor print ISSN are valid']
+
+        self.assertEquals(
+            self._validate("<root><issn pub-type='ppub'>01ols0-OIN</issn></root>"), expected)
+
+    def test_pipe_issn_with_one_strange_ISSN_and_one_valid_ISSN(self):
+        expected = ['ok', '']
+
+        self.assertEquals(
+            self._validate("<root><issn pub-type='ppub'>01ols0-OIN</issn><issn pub-type='epub'>1414-431X</issn></root>"), expected)
