+++ balaio/validator.py
@@ -1,22 +1,12 @@
 # coding: utf-8
-import ConfigParser
-import urllib2
-import urllib
 import sys
-import xml.etree.ElementTree as etree
-import json
-from StringIO import StringIO
-
-import plumber
-
-# futurely scieloapi is package
-import scieloapi
 import utils
-from utils import SingletonMixin, Configuration
+import plumber
 import notifier
-from notifier import Request
-from models import Attempt
+import scieloapi
+import xml.etree.ElementTree as etree
 
+from utils import Configuration
 
 config = Configuration.from_env()
 
@@ -25,73 +15,11 @@ STATUS_WARNING = 'w'
 STATUS_ERROR = 'e'
 
 
-def etree_nodes_value(etree, xpath):
-    """
-    Returns text of a given ``xpath`` of ``etree``
-    """
-    return '\n'.join([node.text for node in etree.findall(xpath)])
-
-
-class Manager(object):
-    """
-    Interface for SciELO API
-    """
-    _main = 'MAIN/QUERY?username=USERNAME&api_key=API_KEY&format=json'
-
-    def __init__(self, api_url='http:/????', username='', api_key=''):
-        super(Manager, self).__init__()
-        self.api_params['USERNAME'] = username
-        self.api_params['API_KEY'] = api_key
-        self.api_params['MAIN'] = api_url
-
-        for key, param in self.api_params.items():
-            self._main = self._main.replace(key, param)
-
-    def do_query(self, query, params={}):
-        """
-        Consulta SciLO Manager API
-        Returns JSON
-        """
-        try:
-            r = urllib2.open(self._main.replace('QUERY', query) + '&'.join([key + '=' + value for key, value in params.items()])).read()
-        except:
-            r = '{}'
-        return json.load(StringIO(r))
-
-    def _item_id(self, query, data_label, match_value):
-        """
-        Find in all an item which ``data_label`` has a value that matches ``match_value``
-        (esse metodo seria desnecessario se na api estivesse search)
-        Returns item id
-        """
-        item_id = None
-        all_items = json.load(self.do_query(query))
-
-        meta = all_items.get('meta', {})
-        total = meta.get('total_count', 0)
-        offset = meta.get('offset', 0)
-        limit = meta.get('limit', 0)
-
-        found = [o for o in all_items.get('objects', {}) if o.get(data_label, '') == match_value]
-        while found is [] and offset < total:
-            offset += limit
-            all_items = json.load(self.do_query(query, {'offset': offset}))
-            found = [o for o in all_items.get('objects', {}) if o.get(data_label, '') == match_value]
-
-        if not found is []:
-            item_id = found[0].get('id', None)
-        return item_id
-
-    def journal(self, value, attribute='id'):
-        item_id = value if attribute == 'id' else self._item_id('journals', attribute, value)
-        return self.do_query('journals/' + item_id + '/')
-
-
 class ValidationPipe(plumber.Pipe):
     """
     Specialized Pipe which validates the data and notifies the result
     """
-    def __init__(self, data, manager_dep=Manager, notifier_dep=notifier.Notifier):
+    def __init__(self, data, manager_dep=scieloapi.Manager, notifier_dep=notifier.Notifier):
         super(ValidationPipe, self).__init__(data)
         self._notifier = notifier_dep()
         self._manager = manager_dep()
@@ -115,28 +43,28 @@ class ValidationPipe(plumber.Pipe):
 
     def compare_registered_data_and_xml_data(self, package_analyzer):
         """
-        Compare registered data in Manager to data in XML
+        Compare registered data in scieloapi.Manager to data in XML
         Returns [status, description]
         """
         registered_data = self._registered_data(package_analyzer)
         xml_data = self._xml_data(package_analyzer)
 
         if registered_data is None:
-            status, description = [STATUS_ERROR, self._registered_data_label + ' not found in Manager']
+            status, description = [STATUS_ERROR, self._registered_data_label + ' not found in scieloapi.Manager']
         elif xml_data == '':
             status, description = [STATUS_ERROR, self._xml_data_label + ' not found in XML']
         elif xml_data == registered_data:
             status, description = [STATUS_OK, xml_data]
         else:
             status = STATUS_ERROR
-            description = 'Data in XML and Manager do not match.' + '\n' + 'Data in Manager: ' + registered_data + '\n' + 'Data in XML: ' + xml_data
+            description = 'Data in XML and scieloapi.Manager do not match.' + '\n' + 'Data in scieloapi.Manager: ' + registered_data + '\n' + 'Data in XML: ' + xml_data
         return [status, description]
 
 
 # Pipes to validate journal data
 class AbbrevJournalTitleValidationPipe(ValidationPipe):
     """
-    Check if journal-meta/abbrev-journal-title[@abbrev-type='publisher'] is the same as registered in Manager
+    Check if journal-meta/abbrev-journal-title[@abbrev-type='publisher'] is the same as registered in scieloapi.Manager
     """
     def validate(self, package_analyzer):
         self._registered_data_label = 'title_iso'
@@ -144,7 +72,7 @@ class AbbrevJournalTitleValidationPipe(ValidationPipe):
         return self.compare_registered_data_and_xml_data(package_analyzer)
 
     def _xml_data(self, package_analyzer):
-        return etree_nodes_value(package_analyzer.xml, self._xml_data_label)
+        return scieloapi.etree_nodes_value(package_analyzer.xml, self._xml_data_label)
 
     def _registered_data(self, package_analyzer):
         return self._manager.journal(package_analyzer.meta['journal_title'], 'title').get(self._registered_data_label, None)
@@ -185,9 +113,10 @@ class FundingCheckingPipe(ValidationPipe):
         # if text contains any number
         return any((True for n in xrange(10) if str(n) in text))
 
+
 class ISSNCheckingPipe(ValidationPipe):
     """
-    Verify if the ISSN(s) exist on SciELO Manager and if it`s a valid ISSN.
+    Verify if the ISSN(s) exist on SciELO scieloapi.Manager and if it`s a valid ISSN.
 
     Analyzed atribute from ``.//issn``: ``@pub-type="ppub"`` or ``@pub-type="epub"``
 
@@ -201,16 +130,16 @@ class ISSNCheckingPipe(ValidationPipe):
 
         if utils.is_valid_issn(journal_pissn) or utils.is_valid_issn(journal_eissn):
             if journal_pissn:
-                #Validate journal_pissn against SciELO Manager
+                #Validate journal_pissn against SciELO scieloapi.Manager
                 pass
             if journal_eissn:
-                #Validate journal_eissn against SciELO Manager
+                #Validate journal_eissn against SciELO scieloapi.Manager
                 pass
             return [STATUS_OK, '']
         else:
             return [STATUS_ERROR, 'neither eletronic ISSN nor print ISSN are valid']
 
-ppl = plumber.Pipeline(FundingCheckingPipe, ISSNCheckingPipe)
+ppl = plumber.Pipeline(ISSNCheckingPipe, FundingCheckingPipe)
 
 if __name__ == '__main__':
     messages = utils.recv_messages(sys.stdin, utils.make_digest)
